Use SIGKILL when restarting services in serviceProxy.
As mentioned in https://ghc.haskell.org/trac/ghc/ticket/4162 GHC API
replaces signal handlers, and therefore SIGTERM receive when test is
running will be caught by HINT and reported back as RuntimeError in
test. As there is no clear work-around for this, we will not use SIGTERM
for service restarts.